+++ CHANGELOG.rst
@@ -3973,7 +3973,7 @@ v0.5.5
 * ``$XONSHRC`` and related configuration variables now accept JSON-based
   static configuration file names as elements. This unifies the two methods
   of run control to a single entry point and loading system.
-* The ``xonsh.shell.Shell()`` class now requires that an Execer instance
+* The ``xonsh.shells.Shell()`` class now requires that an Execer instance
   be explicitly provided to its init method. This class is no longer
   responsible for creating an execer an its dependencies.
 * Moved decorators ``unthreadable``, ``uncapturable`` from